Исправлены замечания после ревью. - Добавлен RestController, возвращающий JSON.

+++ server/src/main/java/ru/paramonov/project/server/common/Card.java
@@ -1,11 +1,11 @@
 package ru.paramonov.project.server.common;
 
 public interface Card {
-    int cardNumber = 0;
+    long cardNumber = 0;
     int personalIdentificationNumber = 0;
     String bankName = null;
 
-    int getCardNumber();
+    long getCardNumber();
     long getAccountNumber();
     int getPersonalIdentificationNumber();
     String getBankName();
